Feat: 분산 락 @DistributedLock 어노테이션 생성 후 관리해라 - #118

+++ tictoc-common/src/main/java/tictoc/constants/RedisConstants.java
@@ -8,6 +8,6 @@ public class RedisConstants {
     public static final String REDIS_KEY_SPACE_EVENT_VALUE = "Ex";
 
     public static final String LOCK_PREFIX = "bid:";
-    public static final double LOCK_WAIT_TIME = 0;
+    public static final long LOCK_WAIT_TIME = 0;
     public static final long LOCK_LEASE_TIME = 3;
 }
